move _typing.py into clorm.orm
