--- clorm/_typing.py
@@ -1,16 +1,0 @@
-
-from typing import Any, Tuple, TypeVar
-
-
-_TP = TypeVar("_TP", bound=Tuple[Any, ...])
-
-_T0 = TypeVar("_T0", bound=Any)
-_T1 = TypeVar("_T1", bound=Any)
-_T2 = TypeVar("_T2", bound=Any)
-_T3 = TypeVar("_T3", bound=Any)
-_T4 = TypeVar("_T4", bound=Any)
-_T5 = TypeVar("_T5", bound=Any)
-_T6 = TypeVar("_T6", bound=Any)
-_T7 = TypeVar("_T7", bound=Any)
-_T8 = TypeVar("_T8", bound=Any)
-_T9 = TypeVar("_T9", bound=Any)
